TryGetValueFromSource added to figure when does the value exist
